configs: New config in drivers/net/can for v5.4-rc1
Message-id: <20191014162300.16308-9-jcline@redhat.com>
Patchwork-id: 278874
O-Subject: [ARK INTERNAL PATCH] configs: New config in drivers/net/can for v5.4-rc1
Bugzilla:
RH-Acked-by: John Linville <linville@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* drivers/net/can/sja1000/Kconfig (help): CAN_F81601 (tristate)
    This driver adds support for Fintek F81601 PCIE to 2 CAN
    Controller. It had internal 24MHz clock source, but it can
    be changed by manufacturer. Use modinfo to get usage for
    parameters. Visit http://www.fintek.com.tw to get more
    information.
* drivers/net/can/Kconfig (help): CAN_KVASER_PCIEFD (tristate)
    This is a driver for the Kvaser PCI Express CAN FD family.

    Supported devices:
      Kvaser PCIEcan 4xHS
      Kvaser PCIEcan 2xHS v2
      Kvaser PCIEcan HS v2
      Kvaser Mini PCI Express HS v2
      Kvaser Mini PCI Express 2xHS v2

Signed-off-by: Jeremy Cline <jcline@redhat.com>