[redhat] New configs in drivers/firmware
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_APPLE_PROPERTIES:

 Retrieve properties from EFI on Apple Macs and assign them to
 devices, allowing for improved support of Apple hardware.
 Properties that would otherwise be missing include the
 Thunderbolt Device ROM and GPU configuration data.

 If unsure, say Y if you have a Mac.  Otherwise N.

 Symbol: APPLE_PROPERTIES [=n]
 Type  : bool
 Defined at drivers/firmware/efi/Kconfig:178
   Prompt: Apple Device Properties
   Depends on: EFI [=y] && EFI_STUB [=y] && X86 [=y]
   Location:
     -> Firmware Drivers
       -> EFI (Extensible Firmware Interface) Support
 Selects: EFI_DEV_PATH_PARSER [=n] && UCS2_STRING [=y]
 Selected by [n]:
   - USB4 [=n] && PCI [=y] && (X86 [=y] || COMPILE_TEST [=n]) && EFI_STUB [=y] && X86 [=y]

---

This is required to be on for CONFIG_USB4=m. As Jarod has requested
that, this has been set to y.

Cc: Jarod Wilson <jarod@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>