[redhat] New configs in drivers/input
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_KEYBOARD_BCM:

 Say Y here if you want to use Broadcom keypad.

 To compile this driver as a module, choose M here: the
 module will be called bcm-keypad.

 Symbol: KEYBOARD_BCM [=n]
 Type  : tristate
 Defined at drivers/input/keyboard/Kconfig:774
   Prompt: Broadcom keypad driver
   Depends on: !UML && INPUT [=y] && INPUT_KEYBOARD [=y] && OF [=y] && HAVE_CLK [=y]
   Location:
     -> Device Drivers
       -> Input device support
         -> Generic input layer (needed for keyboard, mouse, ...) (INPUT [=y])
           -> Keyboards (INPUT_KEYBOARD [=y])
 Selects: INPUT_MATRIXKMAP [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>