[redhat] New configs in net/bridge
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BRIDGE_CFM:

 If you say Y here, then the Ethernet bridge will be able to run CFM
 protocol according to 802.1Q section 12.14

 Say N to exclude this support and reduce the binary size.

 If unsure, say N.

 Symbol: BRIDGE_CFM [=n]
 Type  : bool
 Defined at net/bridge/Kconfig:77
   Prompt: CFM protocol
   Depends on: NET [=y] && BRIDGE [=m]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> 802.1d Ethernet Bridging (BRIDGE [=m])

---

Cc: Ivan Vecera <ivecera@redhat.com>
Cc: Jiri Benc <jbenc@redhat.com>
Cc: Marcelo Leitner <mleitner@redhat.com>
Cc: Davide Caratti <dcaratti@redhat.com>
Cc: Eric Garver <egarver@redhat.com>
Cc: Flavio Leitner <fbl@redhat.com>
Cc: Florian Westphal <fwestpha@redhat.com>
Cc: Guillaume Nault <gnault@redhat.com>
Cc: Hangbin Liu <haliu@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Cc: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Cc: Paolo Abeni <pabeni@redhat.com>
Cc: Phil Sutter <psutter@redhat.com>
Cc: Sabrina Dubroca <sdubroca@redhat.com>
Cc: Stefano Brivio <sbrivio@redhat.com>
Cc: Xin Long <lxin@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>