[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DEBUG_IRQFLAGS:

 Enables checks for potentially unsafe enabling or disabling of
 interrupts, such as calling raw_local_irq_restore() when interrupts
 are enabled.

 Symbol: DEBUG_IRQFLAGS [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:1427
   Prompt: Debug IRQ flag manipulation
   Location:
     -> Kernel hacking
 Selected by [n]:
   - DEBUG_LOCKDEP [=n] && DEBUG_KERNEL [=y] && LOCKDEP [=n]

---

Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>