[redhat] New configs in drivers/crypto
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CRYPTO_DEV_SL3516:

 This option allows you to have support for SL3516 crypto offloader.

 Symbol: CRYPTO_DEV_SL3516 [=n]
 Type  : tristate
 Defined at drivers/crypto/Kconfig:269
   Prompt: Stormlink SL3516 crypto offloader
   Depends on: CRYPTO [=y] && CRYPTO_HW [=y] && HAS_IOMEM [=y] && PM [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Hardware crypto devices (CRYPTO_HW [=y])
 Selects: CRYPTO_SKCIPHER [=y] && CRYPTO_ENGINE [=n] && CRYPTO_ECB [=y] && CRYPTO_AES [=y] && HW_RANDOM [=y]

---

Cc: Herbert Xu <herbert.xu@redhat.com>
Cc: "David S. Miller" <davem@redhat.com>
Cc: Ondrej Mosnacek <omosnace@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>