[redhat] New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ANON_VMA_NAME [=n]
 Type  : bool
 Defined at mm/Kconfig:903
   Prompt: Anonymous VMA name support
   Depends on: PROC_FS [=y] && ADVISE_SYSCALLS [=y] && MMU [=y]
   Location:
     Main menu
       -> Memory Management options

---

Cc: Rafael Aquini <aquini@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Artem Savkov <asavkov@redhat.com>
Cc: Baoquan He <bhe@redhat.com>
Cc: Christoph von Recklinghausen <crecklin@redhat.com>
Cc: Don Dutile <ddutile@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Nico Pache <npache@redhat.com>
Cc: Waiman Long <longman@redhat.com>
Cc: mm-maint@redhat.com
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>