rust: kernel: clean Rust 1.66.0 `rustdoc::broken_intra_doc_links` warnings
Since Rust 1.63.0, `rustdoc` complains with `broken_intra_doc_links`
about intra-doc links pointing to exported `macro_rules`, e.g.:

    error: unresolved link to `dev_info`
       --> rust/kernel/device.rs:135:43
        |
    135 |     /// More details are available from [`dev_info`].
        |                                           ^^^^^^^^ no item named `dev_info` in scope
        |
        = note: `macro_rules` named `dev_info` exists in this crate, but it is not in scope at this link's location
        = note: `-D rustdoc::broken-intra-doc-links` implied by `-D warnings`

    error: aborting due to previous error

The text is confusing, because the link still gets generated, and previous versions
(<= 1.62) did not warn and also generated the link. This was reported
upstream at [1], and it turns out that the link still being generated was
a compatibility measure for docs.rs, which may get removed soon. Thus
the intended behavior is that the user specifies the proper path.

Therefore, clean up the `allow()`s introduced earlier to satisfy `rustdoc`
and the new behavior.

Link: https://github.com/rust-lang/rust/issues/106142 [1]
Reviewed-by: Björn Roy Baron <bjorn3_gh@protonmail.com>
Reviewed-by: Martin Rodriguez Reboredo <yakoyoku@gmail.com>
Tested-by: Martin Rodriguez Reboredo <yakoyoku@gmail.com>
Reviewed-by: Gary Guo <gary@garyguo.net>
Reviewed-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>